refactor: rename checker to output_checker
