+++ src/core/check.rs
@@ -1,1 +1,1 @@
-pub mod checker;
+pub mod output_checker;
