+++ src/core/check/output_checker.rs
